fix(model): store Step4 DSA summaries as widened bf16 on NPU
aclnnIndex (advanced indexing) does not support the fp8_e4m3 dtype on
Ascend, so gathering summaries from the fp8 sidecar cache failed. Store
the e4m3-rounded summary widened to bf16 instead: the widening is
lossless (e4m3 keeps 3 mantissa bits, bf16 keeps 7), so the values are
bit-identical to storing fp8 and casting on read, while every indexing
op stays on supported dtypes.

+++ tests/ut/models/step4/test_step4_dsa.py
@@ -202,8 +202,12 @@ def test_sparse_attention_empty_selection_is_zero():
 def test_gather_region_summaries_follows_block_table():
     torch.manual_seed(5)
     regions_per_block, d = 4, 8
-    # 3 physical blocks x 4 regions each; block 1 is the request's second page.
-    cache = torch.randn(3 * regions_per_block, 1, d).to(torch.float8_e4m3fn)
+    # bf16 cache of e4m3-rounded summaries (the NPU-side storage format).
+    cache = (
+        torch.randn(3 * regions_per_block, 1, d)
+        .to(torch.float8_e4m3fn)
+        .to(torch.bfloat16)
+    )
     block_table = torch.tensor([[0, 2]])
     got = _gather_region_summaries(cache, block_table[0], 8, regions_per_block)
     expected = torch.cat(
